+++ buildacomputer/Models/l_computer_cases_cooling_fans.cs
@@ -7,11 +7,11 @@ namespace buildacomputer.Models
 {
     public class l_computer_cases_cooling_fans
     {
-        public int computer_case_id { get; set; }
+        public long computer_case_id { get; set; }
 
-        public int cooling_fan_size_id { get; set; }
+        public long cooling_fan_size_id { get; set; }
 
-        public int cooling_fan_size_count { get; set; }
+        public long cooling_fan_size_count { get; set; }
 
         public virtual computer_cases computer_cases { get; set; }
         public virtual cooling_fan_sizes cooling_fan_sizes { get; set; }
