+++ buildacomputer/Models/l_motherboard_peripheral_interfaces.cs
@@ -7,11 +7,11 @@ namespace buildacomputer.Models
 {
     public class l_motherboard_peripheral_interfaces
     {
-        public int motherboard_id { get; set; }
+        public long motherboard_id { get; set; }
 
-        public int peripheral_interface_id { get; set; }
+        public long peripheral_interface_id { get; set; }
 
-        public int peripheral_interface_count { get; set; }
+        public long peripheral_interface_count { get; set; }
 
         public virtual motherboards motherboards { get; set; }
         public virtual peripheral_interfaces peripheral_interfaces { get; set; }
