+++ buildacomputer/Models/l_optical_drives_optical_drive_features.cs
@@ -7,9 +7,9 @@ namespace buildacomputer.Models
 {
     public class l_optical_drives_optical_drive_features
     {
-        public int optical_drive_id { get; set; }
+        public long optical_drive_id { get; set; }
 
-        public int optical_drive_feature_id { get; set; }
+        public long optical_drive_feature_id { get; set; }
 
         public virtual optical_drive_features optical_drive_features { get; set; }
         public virtual optical_drives optical_drives { get; set; }
