+++ buildacomputer/Models/power_supplies.cs
@@ -7,21 +7,21 @@ namespace buildacomputer.Models
 {
     public class power_supplies
     {
-        public int power_supply_id { get; set; }
+        public long power_supply_id { get; set; }
 
         public string power_supply_name { get; set; }
 
-        public int total_power_w { get; set; }
+        public long total_power_w { get; set; }
 
-        public int energy_efficiency_percentage { get; set; }
+        public long energy_efficiency_percentage { get; set; }
 
-        public int manufacturer_id { get; set; }
+        public long manufacturer_id { get; set; }
 
-        public int motherboard_form_factor_id { get; set; }
+        public long motherboard_form_factor_id { get; set; }
 
-        public int power_supply_standard_id { get; set; }
+        public long power_supply_standard_id { get; set; }
 
-        public int cooling_fan_size_id { get; set; }
+        public long cooling_fan_size_id { get; set; }
 
         public virtual manufacturers manufacturers { get; set; }
         public virtual motherboard_form_factors motherboard_for_factors { get; set; }
